feat(api): use feature gates
