+++ api/src/lib.rs
@@ -1,7 +1,9 @@
 mod_use::mod_use![utils];
 
+pub mod rpc;
+
 #[cfg(feature = "client")]
 pub mod client;
-pub mod rpc;
+
 #[cfg(feature = "server")]
 pub mod server;
