+++ api/src/rpc/ext.rs
@@ -1,9 +1,12 @@
 //! Contains several extensive traits for core models;
 
-use mongodb::bson;
+use mongodb::bson::Uuid;
 use sg_core::models::{Task, User};
 
-use crate::rpc::{map, models::AddTask, ApiError, ApiResult};
+use crate::{
+    map,
+    rpc::{models::AddTask, ApiError, ApiResult},
+};
 
 pub trait UserExt: Sized {
     fn assert_admin(self) -> ApiResult<Self>;
@@ -20,34 +23,34 @@ impl UserExt for User {
 }
 
 pub trait TaskExt: Sized {
-    fn new_youtube(channel_id: impl Into<String>, parent: mongodb::bson::Uuid) -> Self;
-    fn new_bilibili(uid: impl Into<String>, parent: mongodb::bson::Uuid) -> Self;
-    fn new_twitter(id: impl Into<String>, parent: mongodb::bson::Uuid) -> Self;
+    fn new_youtube(channel_id: impl Into<String>, parent: Uuid) -> Self;
+    fn new_bilibili(uid: impl Into<String>, parent: Uuid) -> Self;
+    fn new_twitter(id: impl Into<String>, parent: Uuid) -> Self;
 }
 
 impl TaskExt for Task {
-    fn new_youtube(channel_id: impl Into<String>, parent: mongodb::bson::Uuid) -> Self {
+    fn new_youtube(channel_id: impl Into<String>, parent: Uuid) -> Self {
         let channel_id = channel_id.into();
         Self {
-            id: bson::Uuid::new(),
+            id: Uuid::new(),
             entity: parent,
             kind: "youtube".to_string(),
             params: map("channel_id", channel_id),
         }
     }
 
-    fn new_bilibili(uid: impl Into<String>, parent: mongodb::bson::Uuid) -> Self {
+    fn new_bilibili(uid: impl Into<String>, parent: Uuid) -> Self {
         Self {
-            id: bson::Uuid::new(),
+            id: Uuid::new(),
             entity: parent,
             kind: "bililive".to_string(),
             params: map("uid", uid),
         }
     }
 
-    fn new_twitter(id: impl Into<String>, parent: mongodb::bson::Uuid) -> Self {
+    fn new_twitter(id: impl Into<String>, parent: Uuid) -> Self {
         Self {
-            id: bson::Uuid::new(),
+            id: Uuid::new(),
             entity: parent,
             kind: "twitter".to_string(),
             params: map("id", id),
